+++ crates/core/src/rules/mod.rs
@@ -9,4 +9,5 @@ pub mod color;
 pub mod damage;
 pub mod direction;
 pub mod items;
+pub mod message;
 pub mod monster_table;